+++ src/main/java/com/programmers/bucketback/domains/vote/api/dto/response/VoteGetByCursorResponse.java
@@ -2,8 +2,8 @@ package com.programmers.bucketback.domains.vote.api.dto.response;
 
 import java.util.List;
 
-import com.programmers.bucketback.domains.vote.application.VoteCursorSummary;
 import com.programmers.bucketback.domains.vote.application.dto.response.GetVotesServiceResponse;
+import com.programmers.bucketback.domains.vote.application.dto.response.VoteCursorSummary;
 
 public record VoteGetByCursorResponse(
 	String nextCursorId,
