+++ src/main/java/com/programmers/bucketback/domains/vote/application/VoteReader.java
@@ -13,8 +13,13 @@ import com.programmers.bucketback.domains.common.vo.CursorPageParameters;
 import com.programmers.bucketback.domains.item.application.ItemReader;
 import com.programmers.bucketback.domains.item.application.vo.ItemInfo;
 import com.programmers.bucketback.domains.item.domain.Item;
+import com.programmers.bucketback.domains.vote.application.dto.request.VoteSortCondition;
+import com.programmers.bucketback.domains.vote.application.dto.request.VoteStatusCondition;
 import com.programmers.bucketback.domains.vote.application.dto.response.GetVoteServiceResponse;
 import com.programmers.bucketback.domains.vote.application.dto.response.GetVotesServiceResponse;
+import com.programmers.bucketback.domains.vote.application.dto.response.VoteCursorSummary;
+import com.programmers.bucketback.domains.vote.application.dto.response.VoteInfo;
+import com.programmers.bucketback.domains.vote.application.dto.response.VoteSummary;
 import com.programmers.bucketback.domains.vote.domain.Vote;
 import com.programmers.bucketback.domains.vote.domain.Voter;
 import com.programmers.bucketback.domains.vote.repository.VoteRepository;
