--- src/main/java/com/programmers/bucketback/domains/vote/application/VoteSortCondition.java
@@ -1,6 +1,0 @@
-package com.programmers.bucketback.domains.vote.application;
-
-public enum VoteSortCondition {
-	POPULARITY,
-	RECENT
-}
