+++ src/main/java/com/programmers/bucketback/domains/vote/application/dto/request/VoteSortCondition.java
@@ -1,0 +1,6 @@
+package com.programmers.bucketback.domains.vote.application.dto.request;
+
+public enum VoteSortCondition {
+	POPULARITY,
+	RECENT
+}
