+++ src/main/java/com/programmers/bucketback/domains/vote/application/dto/response/GetVotesServiceResponse.java
@@ -2,8 +2,6 @@ package com.programmers.bucketback.domains.vote.application.dto.response;
 
 import java.util.List;
 
-import com.programmers.bucketback.domains.vote.application.VoteCursorSummary;
-
 public record GetVotesServiceResponse(
 	String nextCursorId,
 	List<VoteCursorSummary> voteCursorSummaries
