+++ src/main/java/com/programmers/bucketback/domains/vote/application/dto/response/VoteCursorSummary.java
@@ -1,4 +1,4 @@
-package com.programmers.bucketback.domains.vote.application;
+package com.programmers.bucketback.domains.vote.application.dto.response;
 
 import com.programmers.bucketback.domains.item.application.vo.ItemInfo;
 import com.programmers.bucketback.domains.item.domain.Item;
