+++ src/main/java/com/programmers/bucketback/domains/vote/application/dto/response/VoteSummary.java
@@ -1,4 +1,4 @@
-package com.programmers.bucketback.domains.vote.application;
+package com.programmers.bucketback.domains.vote.application.dto.response;
 
 public record VoteSummary(
 	VoteInfo voteInfo,
