+++ src/main/java/com/programmers/bucketback/domains/vote/repository/VoteRepositoryForCursor.java
@@ -3,9 +3,9 @@ package com.programmers.bucketback.domains.vote.repository;
 import java.util.List;
 
 import com.programmers.bucketback.domains.common.Hobby;
-import com.programmers.bucketback.domains.vote.application.VoteSortCondition;
-import com.programmers.bucketback.domains.vote.application.VoteStatusCondition;
-import com.programmers.bucketback.domains.vote.application.VoteSummary;
+import com.programmers.bucketback.domains.vote.application.dto.request.VoteSortCondition;
+import com.programmers.bucketback.domains.vote.application.dto.request.VoteStatusCondition;
+import com.programmers.bucketback.domains.vote.application.dto.response.VoteSummary;
 
 public interface VoteRepositoryForCursor {
 	List<VoteSummary> findAllByCursor(
